Update delete student command names (#67)
