+++ src/main/java/seedu/address/logic/commands/deletestudentcommands/DeleteStudentByIndexCommand.java
@@ -1,41 +1,41 @@
-package seedu.address.logic.commands.deletestudentcommand;
+package seedu.address.logic.commands.deletestudentcommands;
 
 import static java.util.Objects.requireNonNull;
 
+import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.person.Email;
 import seedu.address.model.person.Person;
 
 /**
- * Deletes a student identified using email from the address book.
+ * Deletes a student identified using index from the address book.
  */
-public class DeleteStudentByEmailCommand extends DeleteStudentCommand {
+public class DeleteStudentByIndexCommand extends DeleteStudentCommand {
 
-    public static final String MESSAGE_PERSON_EMAIL_NOT_FOUND = "The student with email %s "
+    public static final String MESSAGE_PERSON_INDEX_NOT_FOUND = "The student at index %s "
             + "does not exist in the address book";
 
-    private final Email email;
+    private final Index targetIndex;
 
     /**
      * Creates a DeleteStudentCommand to delete the student with the specified
      * {@code studentId}.
      */
-    public DeleteStudentByEmailCommand(Email email) {
-        this.email = email;
+    public DeleteStudentByIndexCommand(Index targetIndex) {
+        this.targetIndex = targetIndex;
     }
 
     @Override
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
         Person personToDelete;
-        personToDelete = model.searchPersonByPredicate(person -> person.getEmail().equals(email));
-        if (personToDelete == null) {
-            throw new CommandException(String.format(MESSAGE_PERSON_EMAIL_NOT_FOUND, email));
-
+        try {
+            personToDelete = model.getFilteredPersonList().get(targetIndex.getZeroBased());
+        } catch (IndexOutOfBoundsException e) {
+            throw new CommandException(String.format(MESSAGE_PERSON_INDEX_NOT_FOUND, targetIndex.getOneBased()));
         }
         model.deletePerson(personToDelete);
 
@@ -49,18 +49,18 @@ public class DeleteStudentByEmailCommand extends DeleteStudentCommand {
         }
 
         // instanceof handles nulls
-        if (!(other instanceof DeleteStudentByEmailCommand)) {
+        if (!(other instanceof DeleteStudentByIndexCommand)) {
             return false;
         }
 
-        DeleteStudentByEmailCommand otherDeleteCommand = (DeleteStudentByEmailCommand) other;
-        return email.equals(otherDeleteCommand.email);
+        DeleteStudentByIndexCommand otherDeleteCommand = (DeleteStudentByIndexCommand) other;
+        return targetIndex.equals(otherDeleteCommand.targetIndex);
     }
 
     @Override
     public String toString() {
         return new ToStringBuilder(this)
-                .add("email", email)
+                .add("targetIndex", targetIndex)
                 .toString();
     }
 }
